Reverts previous commit
Reason:
A user might delete a photographer from the YOURS section of Photographer Manager because it shows that they have 0 from that photographer. This will not only delete the photographer but also all the photos added by other users under that photographer. Maybe it's best to add a count column showing the count of photos from others in this section.
Earlier, before this commit [c436bfa], all photographers with the count of 0 were hidden from the photographer manager's YOURS section, whether they were added by the user or not. Maybe that was the best approach. I can't be sure. Need to find a solution for this. If not, I should just revert the commit instead. It might not have been the best approach but it definitely prevented this problem. But it confuses the user on where the photographer disappeared once they deleted a photo from that photographer, if the photographer was added along with that photo. Before the previous commit, the photographer would not appear in the YOURS section but would be present in the ALL section.
Time to revert the commit.